- Fixed an issue where hitting Attack Speed Upgrade would cause enemies to become stronger (Going past -100% caused this issue) - Fixed an issue where having too low attack speed and holding down Space bar would prevent attacks from happening - Unlockables have been slightly tweaked - Projectile Split unlockable ability proc chance has been fixed - Fixed a visual issue happening when splitting projectiles - Fixed an issue where bouncing ability would cause unwanted/invalid projectile behaviour

+++ Assets/Scripts/Main Menu/UpgradeToolTip.cs
@@ -13,7 +13,7 @@ public class UpgradeToolTip : MonoBehaviour, IPointerEnterHandler, IPointerExitH
     public float NextBonus;
 
     public int UpgradeCost;
-    public int CostIncreasePerUpgrade;
+    public float CostIncreasePerUpgrade;
 
     public int Level;
     private Text textTarget;
@@ -86,12 +86,12 @@ public class UpgradeToolTip : MonoBehaviour, IPointerEnterHandler, IPointerExitH
 
         if (UpgradeCost > 0)
             stats.Currency -= UpgradeCost;
-        UpgradeCost += CostIncreasePerUpgrade;
+        UpgradeCost += (int)CostIncreasePerUpgrade;
 
         if (BonusIncreasePerUpgrade != 1)
             BonusIncreasePerUpgrade = (float)(BonusIncreasePerUpgrade * 0.985);
 
-        CostIncreasePerUpgrade = (int)(CostIncreasePerUpgrade * 1.10);
+        CostIncreasePerUpgrade = (float)(CostIncreasePerUpgrade * 1.10);
         Level++;
 
         PrintToolTip();
